Removing wrapper component and just adding examples on how to use, WIP
